backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
PyJWT
python-multipart
itsdangerous
alembic
aiosqlite
greenlet
passlib[bcrypt]
